Enhance [Internal] [Authentication] [Cryptography] Reorganize Hybrid Cryptosystem
- [+] refactor(crypto): remove unused code and reorganize hybrid stream encryption
- [+] remove unused functions DecryptLargeData, EncryptLargeData, and processLargeData
- [+] remove unused io import
- [+] remove HybridEncryptStream and HybridDecryptStream methods from Service interface
- [+] remove HybridEncryptStream and HybridDecryptStream implementations from cryptoService
- [+] move hybrid stream encryption tests to hybrid/hybrid_test.go and hybrid/stream/stream_test.go
- [+] create new package hybrid/stream for hybrid stream encryption
- [+] rename HybridEncryptStream to EncryptStream and HybridDecryptStream to DecryptStream
- [+] add StreamService interface for stream-based encryption in hybrid package
- [+] implement streamService struct that satisfies StreamService interface
- [+] add tests for StreamService Encrypt and Decrypt methods
