+++ backend/internal/middleware/authentication/crypto/crypto_test.go
@@ -5,7 +5,6 @@
 package crypto_test
 
 import (
-	"bytes"
 	"crypto/rand"
 	"encoding/base64"
 	"h0llyw00dz-template/backend/internal/middleware/authentication/crypto"
@@ -152,131 +151,4 @@ func TestCryptoService(t *testing.T) {
 		t.Error("Expected invalid ciphertext due to corruption, but got valid")
 	}
 
-	// New test case: Hybrid encrypt and decrypt stream
-	t.Run("HybridEncryptDecryptStream", func(t *testing.T) {
-		useArgon2 := false
-		secryptKey := "gopher-testing-testing-testinggg"
-		signKey := "gopher-testing-testing-testing"
-		service := crypto.New(useArgon2, secryptKey, signKey)
-
-		// Generate random keys for AES and ChaCha20-Poly1305.
-		aesKey := make([]byte, 32)    // AES-256 requires a 32-byte key.
-		chachaKey := make([]byte, 32) // ChaCha20-Poly1305 uses a 32-byte key.
-
-		_, err := rand.Read(aesKey)
-		if err != nil {
-			t.Fatalf("Failed to generate AES key: %v", err)
-		}
-
-		_, err = rand.Read(chachaKey)
-		if err != nil {
-			t.Fatalf("Failed to generate ChaCha20-Poly1305 key: %v", err)
-		}
-
-		// Simulate plaintext data to encrypt.
-		plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
-
-		// Encrypt the data using the service.
-		inputBuffer := bytes.NewBuffer(plaintext)
-		encryptedBuffer := new(bytes.Buffer)
-		err = service.HybridEncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
-		if err != nil {
-			t.Fatalf("Failed to encrypt data: %v", err)
-		}
-
-		// Ensure the encrypted data buffer's read position is reset to the beginning.
-		encryptedData := encryptedBuffer.Bytes()
-		encryptedBuffer = bytes.NewBuffer(encryptedData)
-
-		// Decrypt the data using the service.
-		decryptedBuffer := new(bytes.Buffer)
-		err = service.HybridDecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
-		if err != nil {
-			t.Fatalf("Failed to decrypt data: %v", err)
-		}
-
-		// Compare the decrypted data to the original plaintext.
-		decryptedData := decryptedBuffer.Bytes()
-		if !bytes.Equal(decryptedData, plaintext) {
-			t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
-		}
-	})
-}
-
-func TestHybridEncryptDecryptStream(t *testing.T) {
-	// Generate random keys for AES and ChaCha20-Poly1305.
-	aesKey := make([]byte, 32)    // AES-256 requires a 32-byte key.
-	chachaKey := make([]byte, 32) // ChaCha20-Poly1305 uses a 32-byte key.
-
-	_, err := rand.Read(aesKey)
-	if err != nil {
-		t.Fatalf("Failed to generate AES key: %v", err)
-	}
-
-	_, err = rand.Read(chachaKey)
-	if err != nil {
-		t.Fatalf("Failed to generate ChaCha20-Poly1305 key: %v", err)
-	}
-
-	// Simulate plaintext data to encrypt.
-	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
-
-	// Encrypt the data.
-	inputBuffer := bytes.NewBuffer(plaintext)
-	encryptedBuffer := new(bytes.Buffer)
-	err = crypto.HybridEncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
-	if err != nil {
-		t.Fatalf("Failed to encrypt data: %v", err)
-	}
-
-	// Ensure the encrypted data buffer's read position is reset to the beginning.
-	encryptedData := encryptedBuffer.Bytes()
-	encryptedBuffer = bytes.NewBuffer(encryptedData)
-
-	// Decrypt the data.
-	decryptedBuffer := new(bytes.Buffer)
-	err = crypto.HybridDecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
-	if err != nil {
-		t.Fatalf("Failed to decrypt data: %v", err)
-	}
-
-	// Compare the decrypted data to the original plaintext.
-	decryptedData := decryptedBuffer.Bytes()
-	if !bytes.Equal(decryptedData, plaintext) {
-		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
-	}
-}
-
-func TestHybridEncryptDecryptStreamWithApiKey(t *testing.T) {
-	// Predefined API keys or secret keys, which should be securely stored and retrieved.
-	aesKey := []byte("gopher-testing-testing-testinggg")
-	chachaKey := []byte("gopher-testing-testing-testinggg")
-
-	// Simulate plaintext data to encrypt.
-	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
-
-	// Encrypt the data.
-	inputBuffer := bytes.NewBuffer(plaintext)
-	encryptedBuffer := new(bytes.Buffer)
-	err := crypto.HybridEncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
-	if err != nil {
-		t.Fatalf("Failed to encrypt data: %v", err)
-	}
-
-	// Ensure the encrypted data buffer's read position is reset to the beginning.
-	encryptedData := encryptedBuffer.Bytes()
-	encryptedBuffer = bytes.NewBuffer(encryptedData)
-
-	// Decrypt the data.
-	decryptedBuffer := new(bytes.Buffer)
-	err = crypto.HybridDecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
-	if err != nil {
-		t.Fatalf("Failed to decrypt data: %v", err)
-	}
-
-	// Compare the decrypted data to the original plaintext.
-	decryptedData := decryptedBuffer.Bytes()
-	if !bytes.Equal(decryptedData, plaintext) {
-		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
-	}
 }
