+++ backend/internal/middleware/authentication/crypto/hybrid/hybrid_test.go
@@ -5,7 +5,10 @@
 package hybrid_test
 
 import (
+	"bytes"
+	"crypto/rand"
 	"h0llyw00dz-template/backend/internal/middleware/authentication/crypto/hybrid"
+	"h0llyw00dz-template/backend/internal/middleware/authentication/crypto/hybrid/stream"
 	"testing"
 
 	"github.com/gofiber/fiber/v2/middleware/encryptcookie"
@@ -115,4 +118,143 @@ func TestEncryptDecryptCookie(t *testing.T) {
 			t.Errorf("Got: %v", err)
 		}
 	})
+}
+
+func TestStreamEncryptDecrypt(t *testing.T) {
+	// Generate random keys for AES and ChaCha20-Poly1305.
+	aesKey := make([]byte, 32)    // AES-256 requires a 32-byte key.
+	chachaKey := make([]byte, 32) // ChaCha20-Poly1305 uses a 32-byte key.
+
+	_, err := rand.Read(aesKey)
+	if err != nil {
+		t.Fatalf("Failed to generate AES key: %v", err)
+	}
+
+	_, err = rand.Read(chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to generate ChaCha20-Poly1305 key: %v", err)
+	}
+
+	// Test cases
+	testCases := []struct {
+		name  string
+		value string
+	}{
+		{
+			name:  "Simple data value",
+			value: "hello world",
+		},
+		{
+			name:  "Complex data value",
+			value: "!@#$%^&*()_+=-`~[]{}|;':\"<>,.?/\\",
+		},
+		{
+			name:  "Empty data value",
+			value: "",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			// Create an instance of the stream encryption service
+			service := hybrid.NewStreamService(aesKey, chachaKey)
+
+			// Encrypt the data value
+			encryptedData, err := service.Encrypt(tc.value)
+			if err != nil {
+				t.Fatalf("Failed to encrypt data: %v", err)
+			}
+
+			// Decrypt the encrypted data value
+			decryptedValue, err := service.Decrypt(encryptedData)
+			if err != nil {
+				t.Fatalf("Failed to decrypt data: %v", err)
+			}
+
+			// Compare the decrypted value with the original value
+			if decryptedValue != tc.value {
+				t.Errorf("Decrypted value does not match the original value")
+				t.Errorf("Expected: %s", tc.value)
+				t.Errorf("Got: %s", decryptedValue)
+			}
+		})
+	}
+}
+
+func TestHybridEncryptDecryptStream(t *testing.T) {
+	// Generate random keys for AES and ChaCha20-Poly1305.
+	aesKey := make([]byte, 32)    // AES-256 requires a 32-byte key.
+	chachaKey := make([]byte, 32) // ChaCha20-Poly1305 uses a 32-byte key.
+
+	_, err := rand.Read(aesKey)
+	if err != nil {
+		t.Fatalf("Failed to generate AES key: %v", err)
+	}
+
+	_, err = rand.Read(chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to generate ChaCha20-Poly1305 key: %v", err)
+	}
+
+	// Simulate plaintext data to encrypt.
+	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
+
+	// Encrypt the data.
+	inputBuffer := bytes.NewBuffer(plaintext)
+	encryptedBuffer := new(bytes.Buffer)
+	err = stream.EncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to encrypt data: %v", err)
+	}
+
+	// Ensure the encrypted data buffer's read position is reset to the beginning.
+	encryptedData := encryptedBuffer.Bytes()
+	encryptedBuffer = bytes.NewBuffer(encryptedData)
+
+	// Decrypt the data.
+	decryptedBuffer := new(bytes.Buffer)
+	err = stream.DecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to decrypt data: %v", err)
+	}
+
+	// Compare the decrypted data to the original plaintext.
+	decryptedData := decryptedBuffer.Bytes()
+	if !bytes.Equal(decryptedData, plaintext) {
+		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
+	}
+}
+
+func TestHybridEncryptDecryptStreamWithApiKey(t *testing.T) {
+	// Predefined API keys or secret keys, which should be securely stored and retrieved.
+	aesKey := []byte("gopher-testing-testing-testinggg")
+	chachaKey := []byte("gopher-testing-testing-testinggg")
+
+	// Simulate plaintext data to encrypt.
+	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
+
+	// Encrypt the data.
+	inputBuffer := bytes.NewBuffer(plaintext)
+	encryptedBuffer := new(bytes.Buffer)
+	err := stream.EncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to encrypt data: %v", err)
+	}
+
+	// Ensure the encrypted data buffer's read position is reset to the beginning.
+	encryptedData := encryptedBuffer.Bytes()
+	encryptedBuffer = bytes.NewBuffer(encryptedData)
+
+	// Decrypt the data.
+	decryptedBuffer := new(bytes.Buffer)
+	err = stream.DecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to decrypt data: %v", err)
+	}
+
+	// Compare the decrypted data to the original plaintext.
+	decryptedData := decryptedBuffer.Bytes()
+	if !bytes.Equal(decryptedData, plaintext) {
+		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
+	}
 }
