+++ backend/internal/middleware/authentication/crypto/hybrid/stream/stream_test.go
@@ -1,0 +1,91 @@
+// Copyright (c) 2024 H0llyW00dz All rights reserved.
+//
+// License: BSD 3-Clause License
+
+package stream_test
+
+import (
+	"bytes"
+	"crypto/rand"
+	"testing"
+
+	"h0llyw00dz-template/backend/internal/middleware/authentication/crypto/hybrid/stream"
+)
+
+func TestHybridEncryptDecryptStream(t *testing.T) {
+	// Generate random keys for AES and ChaCha20-Poly1305.
+	aesKey := make([]byte, 32)    // AES-256 requires a 32-byte key.
+	chachaKey := make([]byte, 32) // ChaCha20-Poly1305 uses a 32-byte key.
+
+	_, err := rand.Read(aesKey)
+	if err != nil {
+		t.Fatalf("Failed to generate AES key: %v", err)
+	}
+
+	_, err = rand.Read(chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to generate ChaCha20-Poly1305 key: %v", err)
+	}
+
+	// Simulate plaintext data to encrypt.
+	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
+
+	// Encrypt the data.
+	inputBuffer := bytes.NewBuffer(plaintext)
+	encryptedBuffer := new(bytes.Buffer)
+	err = stream.EncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to encrypt data: %v", err)
+	}
+
+	// Ensure the encrypted data buffer's read position is reset to the beginning.
+	encryptedData := encryptedBuffer.Bytes()
+	encryptedBuffer = bytes.NewBuffer(encryptedData)
+
+	// Decrypt the data.
+	decryptedBuffer := new(bytes.Buffer)
+	err = stream.DecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to decrypt data: %v", err)
+	}
+
+	// Compare the decrypted data to the original plaintext.
+	decryptedData := decryptedBuffer.Bytes()
+	if !bytes.Equal(decryptedData, plaintext) {
+		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
+	}
+}
+
+func TestHybridEncryptDecryptStreamWithApiKey(t *testing.T) {
+	// Predefined API keys or secret keys, which should be securely stored and retrieved.
+	aesKey := []byte("gopher-testing-testing-testinggg")
+	chachaKey := []byte("gopher-testing-testing-testinggg")
+
+	// Simulate plaintext data to encrypt.
+	plaintext := []byte("Hello, World! This is a test of the hybrid encryption system.")
+
+	// Encrypt the data.
+	inputBuffer := bytes.NewBuffer(plaintext)
+	encryptedBuffer := new(bytes.Buffer)
+	err := stream.EncryptStream(inputBuffer, encryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to encrypt data: %v", err)
+	}
+
+	// Ensure the encrypted data buffer's read position is reset to the beginning.
+	encryptedData := encryptedBuffer.Bytes()
+	encryptedBuffer = bytes.NewBuffer(encryptedData)
+
+	// Decrypt the data.
+	decryptedBuffer := new(bytes.Buffer)
+	err = stream.DecryptStream(encryptedBuffer, decryptedBuffer, aesKey, chachaKey)
+	if err != nil {
+		t.Fatalf("Failed to decrypt data: %v", err)
+	}
+
+	// Compare the decrypted data to the original plaintext.
+	decryptedData := decryptedBuffer.Bytes()
+	if !bytes.Equal(decryptedData, plaintext) {
+		t.Errorf("Decrypted data does not match original plaintext. Got: %s, Want: %s", decryptedData, plaintext)
+	}
+}
